Add CardViewGroup for Android events
